+++ package.json
@@ -28,6 +28,7 @@
     "dialog-polyfill": "^0.5.6",
     "eslint": "^7.18.0",
     "fa-icons": "^0.2.0",
+    "lit-element": "^2.4.0",
     "npm-run-all": "^4.1.5",
     "pre-commit": "^1.2.2",
     "prettier": "^2.2.1",
