+++ src/index.html
@@ -25,8 +25,7 @@
 	<span id="timer"></span>
 	<div id="players"></div>
 	<canvas height="800px" width="800px"></canvas>
-	<div id="chat"></div>
-	<input id="chatInput" />
+	<app-chat />
 	<dialog></dialog>
 	<div id="links">
 		<a href="https://github.com/RubenNL/drraw/" target="_blank"><img src="/images/github_50px.webp" /></a><br />
@@ -43,9 +42,6 @@
 	canvas {
 		border: 1px solid red;
 		float: left;
-	}
-	#chat {
-		overflow-y: scroll;
 	}
 	#players {
 		float: left;
